+++ HISTORY.rst
@@ -111,6 +111,7 @@
 * Remove the include fontawesome and download it from
   the NPMregistry.
 * Fix issue installing scoped NPM packages.
+* Add new icons classes and templates.
 
 3.1.11 (2019-04-XX)
 ===================